+++ src/atlas/pp/__init__.py
@@ -1,1 +1,4 @@
 from .basic import preprocessing
+from .utils import _safe_mudata
+
+__all__ = ["preprocessing"]